Extract mouse stuff from main.ts
